python-server/requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==24.1.0
cached-property==1.5.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
decorator==5.1.1
docstring_parser==0.16
executing==2.0.1
fastapi==0.112.0
frozenlist==1.4.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.60.1
grpcio-status==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
inflection==0.5.1
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpickle==3.2.2
jsonpointer==3.0.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-experimental==0.0.64
langchain-google-vertexai==1.0.8
langchain-text-splitters==0.2.2
langgraph==0.1.19
langsmith==0.1.96
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
orjson==3.10.6
packaging==24.1
parso==0.8.4
pexpect==4.9.0
portalocker==2.10.1
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psycopg==3.2.1
psycopg-pool==3.2.2
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyvis==0.3.2
PyYAML==6.0.1
qdrant-client==1.10.1
rapidfuzz==3.9.5
requests==2.32.3
rsa==4.9
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlalchemy-orm==1.2.10
stack-data==0.6.3
starlette==0.37.2
tenacity==8.5.0
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
wcwidth==0.2.13
yarl==1.9.4
